Reorganized the code, deleted repeated files
