+++ .vscode/settings.json
@@ -2,6 +2,7 @@
     "files.associations": {
         "print_grid.h": "c",
         "display_grid.h": "c",
-        "time.h": "c"
+        "time.h": "c",
+        "calc_gen.h": "c"
     }
 }
