--- Practice_Guide_11/G1_TP11_E5/Front_End/get_input.h
@@ -1,33 +1,0 @@
-/* ************************************************************
-    * CONFIGURACIÓN DEL HEADER *
-   ************************************************************ */
-
-//Nos aseguramos de que el Header solo se incluya una vez 
-#ifndef GET_INPUT
-#define GET_INPUT
-
-/* --------------- LIBRERÍAS UTILIZADAS --------------- */
-
-// Librería estándar de entrada y salida en C
-#include <stdio.h>
-
-// Librería para trabajar con booleanos (verdadero y falso)
-#include <stdbool.h>
-
-/* --------------- VALIDEZ DE CARACTERES --------------- */
-
-// Definimos una macro para validar el caracter (c) insertado por el usuario
-#define IS_VALID_DIGIT(c) ((c) >= '0' && (c) <= '9')
-
-// Definimos una constante para indicar que el usuario desea salir
-#define EXIT_REQUEST -1
-
-// Declaramos la flag para validar si el usuario insertó un caracter inválido
-extern bool syntax_error;
-
-/* --------------- PROTOTIPOS DE FUNCIONES PÚBLICAS --------------- */
-
-// La función lee los caracteres insertados por el usuario, hasta que presione Enter para confirmar o 'q' (o 'Q') para salir
-int read_user_input(void);
-
-#endif
